requirements.txt
nnsight>=0.0.1
torch>=2.0.0
tqdm>=4.65.0
matplotlib>=3.7.2
ipykernel>=6.25.1
python-dotenv>=1.1.0
transformers>=4.37.0
huggingface-hub>=0.20.0
accelerate>=0.27.0
types-tqdm>=4.67.0
transformer-lens>=0.1.1
plotly>=5.22.0
nbformat>=5.9.0
openai>=1.1.0
pandas>=2.1.0
pandas-stubs>=2.1.0
seaborn>=0.13.0
types-seaborn>=0.13.0